[ui] remove corePlugins preflight false again
It seems that it's okay to include tailwind base in apps now. I couldn't
reproduce the previous problem where the base styles in the app reset
some of the styles defined in UI Components. Reset styles are applied
twice but they happen very early in the chain and don't overwrite
anything.