[ui] DateTimePicker: set Calendar or Clock icon as an inlined svg background-image
Setting the calendar icon (or clock icon for a pure timepicker) as an inlined svg background-image instead of a dedicated Icon element allows us to get rid of the `handleCalendarIconClick` handler which internally called the `open()` method of flatpickr which caused various problems.

The downside is we can not use css-variables for the colors of the icons depending on the theme on inlined svg background-images, as they are not considered part of the DOM.

The solution is to have hardcoded data-url strings for each icon for each theme and apply these via sass. The downside of this solution is we would have to create new data-url strings for each of these icons in case we add a new theme.

In the future we may want to look into a more elegant solution.